Finished remove from cart button
