+++ src/cart/index.html
@@ -21,7 +21,7 @@
         <h2>My Cart</h2>
 
         <ul class="product-list">
-          <!-- Filled with js -->
+          <!-- Filled with cart.js -->
         </ul>
         <div class="cart-footer">
           <p class="cart-total">Total:</p>
