+++ src/js/productDetails.mjs
@@ -10,6 +10,7 @@ export default async function productDetails(productId) {
 }
 
 function renderProductDetails(item) {
+  // Add content to the existing product_page index.html
     document.querySelector("#productName").innerHTML = item.Name;
     document.querySelector("#productNameWithoutBrand").textContent = item.NameWithoutBrand;
     document.querySelector("#productImage").setAttribute("src", item.Image);
